+++ src/main/kotlin/racing/domain/Car.kt
@@ -1,4 +1,6 @@
-package racing
+package racing.domain
+
+import racing.MoveStrategy
 
 data class Car(var distance: Int = 0, val moveStrategy: MoveStrategy, val name: Name) {
     fun move() {
