+++ src/main/kotlin/racing/domain/Name.kt
@@ -1,4 +1,4 @@
-package racing
+package racing.domain
 
 @JvmInline
 value class Name private constructor(val name: String) {
