+++ src/test/kotlin/racing/CarTest.kt
@@ -2,6 +2,8 @@ package racing
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
+import racing.domain.Car
+import racing.domain.Name
 
 class CarTest {
