+++ src/test/kotlin/racing/CarsTest.kt
@@ -4,6 +4,9 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
+import racing.domain.Car
+import racing.domain.Cars
+import racing.domain.Name
 
 class CarsTest {
     private val moveStrategy = MoveStrategy { true }
