+++ src/test/kotlin/racing/NameTest.kt
@@ -3,6 +3,7 @@ package racing
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
+import racing.domain.Name
 
 class NameTest {
